Move storage/* into common
