+++ openverse_catalog/dags/common/storage/util.py
