+++ openverse_catalog/dags/providers/provider_api_scripts/europeana.py
@@ -17,7 +17,7 @@ from datetime import datetime, timedelta, timezone
 from common.licenses import get_license_info
 from common.loader import provider_details as prov
 from common.requester import DelayedRequester
-from storage.image import ImageStore
+from common.storage.image import ImageStore
 
 
 logging.basicConfig(
