+++ openverse_catalog/dags/providers/provider_api_scripts/nypl.py
@@ -5,7 +5,7 @@ from urllib.parse import parse_qs, urlparse
 from common.licenses import get_license_info
 from common.loader import provider_details as prov
 from common.requester import DelayedRequester
-from storage.image import ImageStore
+from common.storage.image import ImageStore
 
 
 logging.basicConfig(
