+++ openverse_catalog/dags/providers/provider_api_scripts/phylopic.py
@@ -15,7 +15,7 @@ import logging
 
 from common.licenses import get_license_info
 from common.requester import DelayedRequester
-from storage.image import ImageStore
+from common.storage.image import ImageStore
 
 
 logging.basicConfig(
