+++ openverse_catalog/dags/providers/provider_api_scripts/raw_pixel.py
@@ -5,7 +5,7 @@ import requests
 from common.licenses import get_license_info
 from common.loader import provider_details as prov
 from common.requester import DelayedRequester
-from storage.image import ImageStore
+from common.storage.image import ImageStore
 
 
 DELAY = 1.0  # time delay (in seconds)
