+++ openverse_catalog/dags/providers/provider_api_scripts/smithsonian.py
@@ -16,7 +16,7 @@ from datetime import datetime
 from common.licenses import LicenseInfo
 from common.loader import provider_details as prov
 from common.requester import DelayedRequester
-from storage.image import ImageStore
+from common.storage.image import ImageStore
 
 
 logger = logging.getLogger(__name__)
