+++ openverse_catalog/dags/providers/provider_api_scripts/walters_art_museum.py
@@ -15,7 +15,7 @@ import os
 
 from common.loader import provider_details as prov
 from common.requester import DelayedRequester
-from storage.image import ImageStore
+from common.storage.image import ImageStore
 
 
 logging.basicConfig(
